+++ src/main/java/dunn/ted/java/instateam/dao/CollaboratorDaoImpl.java
@@ -35,7 +35,7 @@ public class CollaboratorDaoImpl implements CollaboratorDao {
     }
 
     @Override
-    public Collaborator findById(Long id) {
+    public Collaborator findById(int id) {
 
         Session session = sessionFactory.openSession();
