+++ src/main/java/dunn/ted/java/instateam/dao/ProjectDao.java
@@ -9,7 +9,7 @@ import java.util.List;
  */
 public interface ProjectDao {
     List<Project> findAll();
-    Project findById(Long id);
+    Project findById(int id);
     void save(Project project);
     void delete(Project project);
 }
